tools/requirements.txt
lark_parser==0.6.6
Cheetah3==3.1.0
pyelftools==0.26
